* Restore a way to choose the algorithm
git-svn-id: https://scm.gforge.inria.fr/svn/openalea/trunk@1596 04ea39c6-a500-0410-8508-e77fe2a861cf

+++ core/src/core/package.py
@@ -450,8 +450,9 @@ def %s(%s):
             if (src != dst):
                 shutil.copyfile(src, dst)
         except shutil.Error:
-            f = open(dst, 'w')
-            f.close()
+            if not os.path.exists(dst):
+                f = open(dst, 'w')
+                f.close()
 
         newfactory = DataFactory(bname, description)
